Added a call to a shortened url service.
Hi my friend. 
I have added something. Is a new call in case that people that use your package, have a bitly account. Bitly (if you dfont know aobut it) is a shortened url service, so, the call I've added, you can pass the final url to this call, and it will return the shortened url. You need the access token that you generate inside the bitly website.

I've made a test file in my local to test it, and from this final url that I got using imgix:

http://demos.imgix.net/root/Github/tarta1.jpeg?h=100&temp=test&w=100

bitly service return us this one:

http://bit.ly/1Gr8tlQ

Which is the same, but shortened ;)

If you think that this could be helpful for you, then think about it!

Thank you very very much for your patience