refactor: relax validation for max width values
This PR relaxes the constraints on max width values. Prior to this PR
max width values were bounded, now they are not. The only restriction
is that image width values (both max and min) be greater than or equal
to zero (and that the max value is not less than the min value).